Chop old code; Comment recursion; Console writes
Chopped out old code (2x old recursive methods, and their associated
processing/display code).

Comment on the recursive method so it hopefully makes a bit of sense if
I look at it in the future or someone else looks at it.

Added Console.Writes wherever we write to the textBox, just in case you
want it in the console. Which I do